feat: mv waveform to utils
